pyproject.toml
[project]
name = "book_bot"
version = "0.1.0"
description = "book_bot"
requires-python = ">=3.12,<3.14"
dependencies = [
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "pandas-stubs>=2.2.3.241126",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
    "types-psycopg2>=2.9.21.20250121",
    "types-tqdm>=4.67.0.20241221",
]
